requirements.txt
uvicorn==0.32.0
fastapi==0.115.5
python-dotenv==1.0.1
groq==0.12.0
python-multipart==0.0.17
ollama==0.3.3